+++ run_tests.sh
@@ -1,0 +1,4 @@
+#!/usr/bin/env bash
+
+source activate hot-bikes
+python -m unittest discover --pattern="*_tests.py"
